Refactored farms view
