+++ src/components/EarnTriSortAndFilter/EarnTriSortAndFilterSortingType.tsx
@@ -1,0 +1,5 @@
+export enum SortingType {
+  liquidity = 'Liquidity',
+  totalApr = 'Total APR',
+  default = 'Default'
+}
